// src/app/shared/model/codification.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Codification {
    id_codification?: number;
    id?: number;
    n_inventaire: string;
    code_guichet?: string;
    departement?: string;
    n_serie?: string;
    direction?: string;
    famille?: string;
    libelle_famille?: string;
    sous_libelle_famille?: string;
    niveau?: string;
    service?: string;
    sous_famille?: string;
    code_localisation?: string;
    libelle_agence?: string;
    libelle_localisation?: string;
}